feat(fire; close #24): fire burns player
I decided not to set the player on fire because I don't reckon
futuristic firefighting equipment is easily set ablaze.

This also fixes a couples bugs from the player health commit. Noteably
that the bounds on the slider were wrong, and the setter for the current
health wasn't actually updating the current health at all.

Also for good practice made sure to unsubscribe `HealthChanged` from in
`OnDisable`.