fix(#49): hook back up systems
I'm pretty sure in the demo that checkpoints and enemies are broken
because of the refactor changes. I'm not sure if checkpoints are being
registered with the checkpoint system, but I know something I've done
has broken enemies.

Still the primary organization for the refactor should now be done. We
basically just need to take a look at the specific systems and see what
they need.