fix(SceneSystem): eliminate delay in scene registration
Because loading and registration logic for scenes happens in a
couroutine it is not necessarily the case that the `SceneSystem`
registers a scenes info (importantly the `LevelManager`) as soon as it
loads. This is a problem because `Start` and `Awake` run as soon as the
scene load. The solution was to just register the load when objects
request the `LevelManager`.

This commit additionally contains small fixes to get rid of the
duplicate AudioListener spam in the console.