feat(parallax; #83): animated parallax layers
I technically haven't tested this. But I know that it doesn't break any
of the currently setup parallax code so far. What is left now is
scrutinizing the map from UV to texture coords (it's probably going to
be upside down so V likely needs to be flipped), and some slight
alterations to the spacing on the city sprite.