refactor(render; #77): WIP use `RenderGraph` instead of `CommandBuffer`
A bunch of shader issues also got fixed here so the code frankly looks
much nicer. (And as it turns out with the part of Unity's rendering
infrastructure used in ou project you actually can't use `AddRenderPass`
as I was trying to do previously since its apparently legacy.

The code now uses the RenderGraph stuff more heavily so it's actually
less ugly (IMHO).

The silhouette pass works somewhat (it outlines based on the mesh, but
we actually need that to be based on the texture). Otherwise I think the
camera color texture doesn't have an associated stencil buffer cause
when I try to inspect it I get a purely black screen (though I don't
fully understand why the stencil is used in the first place yet).

Otherwise there is most definetly a problem with the JFA Init pass and
so I can't know if the passes after that are working yet. Though I
should be done in a couple hours (famous last words).